% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{dv_dv}
\alias{dv_dv}
\title{Jacobian of a vector-valued function wrt a vector parameter}
\usage{
dv_dv(funct, parm_value, ctx = get_context())
}
\arguments{
\item{funct}{A function \code{f(parm_value, ctx = ctx)}.}

\item{parm_value}{A 1-D numeric or \code{Matrix} vector.}

\item{ctx}{A \code{DerivativeContext}.  Defaults to
\code{\link{get_context}()}.}
}
\value{
A numeric matrix of dimensions
  \code{(length(f(parm_value)), length(parm_value))}.
}
\description{
Returns the full Jacobian matrix for a function that maps a numeric/Matrix
vector to another numeric/Matrix vector.  Both parameter and result must be
1-dimensional (scalar or vector, not a 2-D matrix).
}
