% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.R
\name{get_context}
\alias{get_context}
\title{Retrieve the current derivative context}
\usage{
get_context()
}
\value{
A \code{DerivativeContext} environment.
}
\description{
Returns the \code{DerivativeContext} most recently created by
\code{\link{init_derivative_args}()}.  Raises an error if no context has
been initialised.
}
