% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definedExpression-class.R
\name{is.DefinedExpression}
\alias{is.DefinedExpression}
\title{Test whether an object is a DefinedExpression}
\usage{
is.DefinedExpression(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Logical scalar.
}
\description{
Test whether an object is a DefinedExpression
}
\examples{
is.DefinedExpression("hello")  # FALSE
}
